Fix The Leaderboard Result Record Types And Report Read Failures As Errors
The overall and monthly leaderboard queries have been throwing since the private
result records were introduced. Both records declared `long? SnapshotRank`, copied
across from LeaderboardEntryDto, but the column behind it is an int
([LeagueMemberStats].[SnapshotOverallRank] / [SnapshotMonthRank]). Dapper widens
nothing when it matches a positional record's constructor, so materialisation failed
for every request to /api/leagues/{id}/leaderboard/overall and the monthly equivalent.

Materialising into LeaderboardEntryDto used to work because Dapper maps classes with
init properties by name and converts int to long? on the way; the switch to a
positional record tightened the match to exact CLR types per column. Confirmed the
real column types with sys.sp_describe_first_result_set: SnapshotRank is int in the
overall and monthly queries and bigint in the stage query, whose RANK() window
function makes its long? correct - all three records now carry a comment saying which,
so they are not "harmonised" back into a single shape.

Dapper reports the mismatch as a plain InvalidOperationException, which the API error
middleware maps to 400 Bad Request and logs at Warning because handlers throw that
type for business rules. A server-side defect therefore looked like a client mistake
in the logs, and the 400 body returned the raw Dapper message (including internal type
names) in production. DapperReadDbConnection now translates any InvalidOperationException
out of a read - materialisation mismatches and single-row queries returning multiple
rows - into ReadQueryFailedException, so it lands in the middleware's unhandled bucket
as an Error and a 500.

Adds ThePredictions.Infrastructure.Tests.Unit, the first test project for that
assembly, covering the translation and the exception types that must pass through
untouched. Verified against the dev database by running all three leaderboard handlers
end to end through the real DapperReadDbConnection: the overall handler reproduces the
production error before the type change and materialises after it.

Records the exact-type rule in CLAUDE.md and docs/guides/database.md, the exception
mapping in the API guide, and notes in the test-suite plan that SQLite cannot catch
this class of bug because its type affinities differ from SQL Server's.

+++ src/ThePredictions.Application/Features/Leagues/Queries/GetOverallLeaderboardQueryHandler.cs
@@ -81,12 +81,14 @@ public class GetOverallLeaderboardQueryHandler(
         });
     }
 
+    // SnapshotRank is [LeagueMemberStats].[SnapshotOverallRank], an int column, so it must be int? here
+    // even though LeaderboardEntryDto exposes it as long? - Dapper's constructor match is exact per column.
     [SuppressMessage("ReSharper", "ClassNeverInstantiated.Local")]
     private record OverallLeaderboardQueryResult(
         long Rank,
         string PlayerName,
         int? TotalPoints,
         string UserId,
-        long? SnapshotRank,
+        int? SnapshotRank,
         bool IsRoundInProgress);
 }
